+++ src/command_modules/azure-cli-appservice/tests/test_webapp_commands_thru_mock.py
@@ -21,7 +21,8 @@ from azure.cli.command_modules.appservice.custom import (set_deployment_user,
                                                          _match_host_names_from_cert,
                                                          bind_ssl_cert,
                                                          list_publish_profiles,
-                                                         config_source_control)
+                                                         config_source_control,
+                                                         show_webapp)
 
 # pylint: disable=line-too-long
 from vsts_cd_manager.continuous_delivery_manager import ContinuousDeliveryResult
@@ -214,6 +215,19 @@ class Test_Webapp_Mocked(unittest.TestCase):
         # assert
         webbrowser_mock.assert_called_with('https://haha.com')
         log_mock.assert_called_with('myRG', 'myweb', None, None)
+
+    @mock.patch('azure.cli.command_modules.appservice.custom._generic_site_operation', autospec=True)
+    @mock.patch('azure.cli.command_modules.appservice.custom._rename_server_farm_props', autospec=True)
+    @mock.patch('azure.cli.command_modules.appservice.custom._fill_ftp_publishing_url', autospec=True)
+    def test_show_webapp(self, file_ftp_mock, rename_mock, site_op_mock):
+        faked_web = mock.MagicMock()
+        site_op_mock.return_value = faked_web
+        # action
+        result = show_webapp('myRG', 'myweb', slot=None, app_instance=None)
+        # assert (we invoke the site op)
+        self.assertEqual(faked_web, result)
+        self.assertTrue(rename_mock.called)
+        self.assertTrue(file_ftp_mock.called)
 
     @mock.patch('azure.cli.command_modules.appservice.custom._generic_site_operation', autospec=True)
     def test_sync_repository_skip_bad_error(self, site_op_mock):