+++ src/command_modules/azure-cli-cdn/azure/cli/command_modules/cdn/custom.py
@@ -41,7 +41,6 @@ def update_endpoint(instance,
                     is_https_allowed=None,
                     query_string_caching_behavior=None,
                     tags=None):
-    # pylint: disable=too-many-arguments
     params = EndpointUpdateParameters(
         origin_host_header=origin_host_header,
         origin_path=origin_path,
@@ -77,7 +76,6 @@ def create_endpoint(client, resource_group_name, profile_name, name, origins, lo
                     is_compression_enabled=None, is_http_allowed=None, is_https_allowed=None,
                     query_string_caching_behavior=QueryStringCachingBehavior.ignore_query_string.
                     value, tags=None):
-    # pylint: disable=too-many-arguments
     is_compression_enabled = False if is_compression_enabled is None else is_compression_enabled
     is_http_allowed = True if is_http_allowed is None else is_http_allowed
     is_https_allowed = True if is_https_allowed is None else is_https_allowed
@@ -102,7 +100,6 @@ create_endpoint.__doc__ = Endpoint.__doc__
 
 def create_custom_domain(client, resource_group_name, profile_name, endpoint_name, name, hostname,
                          location=None, tags=None):
-    # pylint: disable=too-many-arguments
     from azure.mgmt.cdn.models import (CustomDomain)
     custom_domain = CustomDomain(location, hostname, tags=tags)
     return client.custom_domains.create(resource_group_name,
@@ -125,7 +122,6 @@ update_profile.__doc__ = ProfileUpdateParameters.__doc__
 def create_profile(client, resource_group_name, name,
                    sku=SkuName.standard_akamai.value,
                    location=None, tags=None):
-    # pylint: disable=too-many-arguments
     from azure.mgmt.cdn.models import (Profile, Sku)
     profile = Profile(location, Sku(name=sku), tags=tags)
     return client.profiles.create(resource_group_name, name, profile)