+++ src/command_modules/azure-cli-vm/tests/test_vm_commands.py
@@ -16,6 +16,8 @@ from azure.cli.testsdk.vcr_test_base import (VCRTestBase,
                                              ResourceGroupVCRTestBase,
                                              JMESPathCheck,
                                              NoneCheck)
+from azure.cli.testsdk import ScenarioTest, ResourceGroupPreparer
+from azure.cli.testsdk import JMESPathCheck as JMESPathCheckV2
 
 TEST_DIR = os.path.abspath(os.path.join(os.path.abspath(__file__), '..'))
 
@@ -274,6 +276,17 @@ class VMGeneralizeScenarioTest(ResourceGroupVCRTestBase):
         self.execute()
 
 
+class VMWindowsLicenseTest(ScenarioTest):
+
+    @ResourceGroupPreparer()
+    def test_windows_vm_license_type(self, resource_group):
+        vm_name = 'winvm'
+        self.cmd('vm create -g {} -n {} --image Win2012R2Datacenter --admin-username clitest1234 --admin-password Test123456789# --license-type Windows_Server'.format(resource_group, vm_name))
+        self.cmd('vm show -g {} -n {}'.format(resource_group, vm_name), checks=[
+            JMESPathCheckV2('licenseType', 'Windows_Server')
+        ])
+
+
 class VMCreateFromUnmanagedDiskTest(ResourceGroupVCRTestBase):
 
     def __init__(self, test_method):
@@ -288,7 +301,10 @@ class VMCreateFromUnmanagedDiskTest(ResourceGroupVCRTestBase):
         vm1 = 'vm1'
         self.cmd('vm create -g {} -n {} --image debian --use-unmanaged-disk --admin-username ubuntu --admin-password testPassword0 --authentication-type password'.format(
             self.resource_group, vm1))
-        vm1_info = self.cmd('vm show -g {} -n {}'.format(self.resource_group, vm1))
+        vm1_info = self.cmd('vm show -g {} -n {}'.format(self.resource_group, vm1), checks=[
+            JMESPathCheck('name', vm1),
+            JMESPathCheck('licenseType', None)
+        ])
         self.cmd('vm stop -g {} -n {}'.format(self.resource_group, vm1))
 
         # import the unmanaged os disk into a specialized managed disk
@@ -841,10 +857,13 @@ def _write_config_file(user_name):
 
 
 class DiagnosticsExtensionInstallTest(ResourceGroupVCRTestBase):
+    """
+    Note that this is currently only for a Linux VM. There's currently no test of this feature for a Windows VM.
+    """
 
     def __init__(self, test_method):
         super(DiagnosticsExtensionInstallTest, self).__init__(__file__, test_method, resource_group='cli_test_vm_vmss_diagnostics_extension')
-        self.storage_account = 'clitestdiagextsa20170227'
+        self.storage_account = 'clitestdiagextsa20170510'
         self.vm = 'testdiagvm'
         self.vmss = 'testdiagvmss'
 
@@ -858,13 +877,12 @@ class DiagnosticsExtensionInstallTest(ResourceGroupVCRTestBase):
         self.execute()
 
     def body(self):
-        storage_key = '123'  # use junk keys, do not retrieve real keys which will get into the recording
+        storage_sastoken = '123'  # use junk keys, do not retrieve real keys which will get into the recording
         _, protected_settings = tempfile.mkstemp()
         with open(protected_settings, 'w') as outfile:
             json.dump({
                 "storageAccountName": self.storage_account,
-                "storageAccountKey": storage_key,
-                "storageAccountEndPoint": "https://{}.blob.core.windows.net/".format(self.storage_account)
+                "storageAccountSasToken": storage_sastoken
             }, outfile)
         protected_settings = protected_settings.replace('\\', '\\\\')
 
@@ -873,14 +891,15 @@ class DiagnosticsExtensionInstallTest(ResourceGroupVCRTestBase):
         template_file = os.path.join(curr_dir, 'sample-public.json').replace('\\', '\\\\')
         with open(template_file) as data_file:
             data = json.load(data_file)
-        data["storageAccount"] = self.storage_account
+        data["StorageAccount"] = self.storage_account
         with open(public_settings, 'w') as outfile:
             json.dump(data, outfile)
         public_settings = public_settings.replace('\\', '\\\\')
 
         checks = [
             JMESPathCheck('virtualMachineProfile.extensionProfile.extensions[0].name', "LinuxDiagnostic"),
-            JMESPathCheck('virtualMachineProfile.extensionProfile.extensions[0].settings.storageAccount', self.storage_account)
+            JMESPathCheck('virtualMachineProfile.extensionProfile.extensions[0].publisher', "Microsoft.Azure.Diagnostics"),
+            JMESPathCheck('virtualMachineProfile.extensionProfile.extensions[0].settings.StorageAccount', self.storage_account)
         ]
 
         self.cmd("vmss diagnostics set -g {} --vmss-name {} --settings {} --protected-settings {}".format(
@@ -891,12 +910,14 @@ class DiagnosticsExtensionInstallTest(ResourceGroupVCRTestBase):
         self.cmd("vm diagnostics set -g {} --vm-name {} --settings {} --protected-settings {}".format(
             self.resource_group, self.vm, public_settings, protected_settings), checks=[
                 JMESPathCheck('name', 'LinuxDiagnostic'),
-                JMESPathCheck('settings.storageAccount', self.storage_account)
+                JMESPathCheck('publisher', 'Microsoft.Azure.Diagnostics'),
+                JMESPathCheck('settings.StorageAccount', self.storage_account)
         ])
 
         self.cmd("vm show -g {} -n {}".format(self.resource_group, self.vm), checks=[
             JMESPathCheck('resources[0].name', 'LinuxDiagnostic'),
-            JMESPathCheck('resources[0].settings.storageAccount', self.storage_account)
+            JMESPathCheck('resources[0].publisher', 'Microsoft.Azure.Diagnostics'),
+            JMESPathCheck('resources[0].settings.StorageAccount', self.storage_account)
         ])