+++ src/command_modules/azure-cli-interactive/HISTORY.rst
@@ -3,6 +3,13 @@
 Release History
 
 
+
+0.3.1 (2017-05-09)
+++++++++++++++++++
+
+* Add link to blog in ‘az interactive —help’ (#3252)
+
+
 0.3.0 (2017-05-05)
 ++++++++++++++++++
 
@@ -10,6 +17,7 @@ Release History
 * Colors Options
 * Rename 'shell' => 'interactive'
 * --progress flag
+
 
 
 0.2.1