+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -1,7 +1,16 @@
 .. :changelog:
-
 Release History
 ===============
+unreleased
+----------
+* diagnostics: Fix incorrect Linux diagnostics default config with update for LAD v.3.0 extension
+* disk: support cross subscription blob import
+* vm: support license type on create
+
+2.0.6 (2017-05-09)
+++++++++++++++++++
+* Minor fixes.
+
 2.0.5 (2017-05-05)
 ++++++++++++++++++
 * avail-set: make UD&FD domain counts optional